Make it unable to delete submitted file after deadline

+++ endpoint/module.py
@@ -197,6 +197,16 @@ class ModuleSubmittedFile(object):
 		submittedFile = _get_submitted_file(id, resp)
 		if submittedFile:
 
+			if session.query(model.SubmittedFile).\
+			filter(model.SubmittedFile.id == submittedFile.id).\
+			join(model.Evaluation, model.Evaluation.id == model.SubmittedFile.evaluation).\
+			join(model.Module, model.Module.id == model.Evaluation.module).\
+			join(model.Task, model.Evaluation.module == model.Module.id).\
+			filter(model.Task.evaluation_public == Frue).\
+			filter(model.Task.time_deadline > datetime.datetime.now()).count() == 0:
+				req.context['result'] = { 'status': 'error', 'error': 'You cannot remove files after deadline' }
+				return
+
 			try:
 				os.remove(submittedFile.path)
